clay/requirements.txt
numpy
sentence-transformers
faiss-cpu
scikit-learn
requests
